chore: move table_dir/region_name/region_id to table crate
