+++ src/datanode/src/lib.rs
@@ -25,5 +25,3 @@ pub mod server;
 pub mod sql;
 #[cfg(test)]
 mod tests;
-
-pub use mito::engine::{region_id, region_name, table_dir};
